Nape Lock Improvements
Now, the camera can be moved up and down while the nape lock is engaged. After killing the target, you had to press the nape lock button twice to lock it again. This was improved by setting the '_napeLock' boolean to false when the target is killed. Some unnecessary assignments removed and repeating code lines merged.